Fixed logic for tracking amount of populated items in array. Passing all unit tests now.
I see my misunderstanding, the size variable is not actually meant to show the maximum capacity of the current array, instead, it is just a variable we use to track the amount of items already populated in the array (not the maximum capacity).

Hence, we should not initialize the size variable as 10, but rather as 0, as the array starts out as being empty (but at the size of 10).

And hence we also want to increment the variable 'size' by one after each item we add to the array.

I modified the code as follows and it now passes all my unit tests